code/requirements.txt
pandas
stable-retro
git@github.com:courtois-neuromod/videogames_tools.git
